Patch out nose2 from install_requires and do not run upstream tests
We do the former since nose2 otherwise would need to be a dependency for the conda package;
I assume the new process still respects dependencies as specified in setup.py.

We do the latter in part to simplify (triggered by changes to how tests are run in rattler-build
and/or by the new v1/v2 format) - the tests are already being run upstream on identical source.